GUI-2370 allow int and float casting on number cols

+++ js/modal/joinModal.js
@@ -989,7 +989,7 @@ window.JoinModal = (function($, JoinModal) {
 
     function contextCheck($table, colNum, type) {
         // only check number and string
-        if (type !== "integer" && type !== "decimal" && type !== "string") {
+        if (type !== "integer" && type !== "float" && type !== "string") {
             return {"max": 0, "min": 0, "total": 0, "variance": 0};
         }
